修复删除目录时,当目录不存在时报错的问题

+++ src/index.ts
@@ -58,6 +58,9 @@ export function copyFolder(source: string, target: string, options: IOptions = {
  * @param path
  */
 export function deleteFolder(path: string) {
+  if (!existsSync(path)) {
+    return;
+  }
   switch (type()) {
     case "Windows_NT": {
       execSync(`rd /S /Q "${path}"`);
